+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.14.0] - 2025-09-01
+### Changed
+- **Quick Alchemy:** Sorting inventory and craft lists alphabetically
+- **Quick Alchemy:** Excluding already unstable and versatile-vials 
 ### Fixed
 - **Quick Alchemy:** fixed bug in Unstable concoction that posted note in chat card about Flat check twice, and completed localization for additions.